requirements.txt
allure-pytest==2.13.5
assertpy==1.1
python-dotenv==1.0.1
pytest==8.3.5
requests==2.32.3
